gr: move hier_block.h methods to .cc
