+++ gr/lib/meson.build
@@ -56,7 +56,8 @@ runtime_sources = [
   'buffer_cpu_vmcirc_mmap_shm_open.cc',
   'buffer_net_zmq.cc',
   'rpc_client_interface.cc',
-  'sptr_magic.cc'
+  'sptr_magic.cc',
+  'hier_block.cc'
 ]
 
 if USE_CUDA
